Adding leaves to trees; block place requests can overflow in X,Y,Z (instead of just Z)

+++ mc_raylib_cpp/include/world.h
@@ -56,7 +56,7 @@ class World : public NoCopy {
 		void generateChunk(PerlinNoise& pn);
 
 		int getHeightAt(PerlinNoise& pn, int x, int z);
-		void setBlockAt(Chunk& chunk, size_t x, size_t y, int z, Block block);
+		void setBlockAt(Chunk& chunk, int x, int y, int z, Block block);
 		void createChunkData(PerlinNoise& pn, Chunk& chunk);
 
 		void updateChunkModels();
